update pdf modals
