+++ apps/pds/src/components/fixed/panels/WorkExpPanel.tsx
@@ -69,6 +69,12 @@ export default function WorkExpPanel(): JSX.Element {
             </div>
           )}
           <WorkExp />
+          {isMobile && workExperience.length > 3 && (
+            <div className="flex w-full gap-1 justify-between pt-6">
+              <SolidPrevButton onClick={onPrev} type="button" />
+              <SolidNextButton onClick={onSubmit} type="button" />
+            </div>
+          )}
         </>
       </Page>
       {!isMobile && (
@@ -76,13 +82,6 @@ export default function WorkExpPanel(): JSX.Element {
           <PrevButton action={onPrev} type="button" />
           <NextButton action={onSubmit} type="button" />
         </>
-      )}
-
-      {isMobile && workExperience.length > 3 && (
-        <div className="flex w-full gap-1 justify-between pt-6">
-          <SolidPrevButton onClick={onPrev} type="button" />
-          <SolidNextButton onClick={onSubmit} type="button" />
-        </div>
       )}
     </>
   );
