+++ apps/portal/src/components/fixed/leaves/LeaveLedgerPdf.tsx
@@ -156,7 +156,7 @@ export const remarksHyphenationCallback = (word: string) => {
 };
 
 export const LeaveLedgerPdf: FunctionComponent<LeaveLedgerPdfProps> = ({ employeeData, leaveLedger }) => {
-  const [isClient, setIsClient] = useState<boolean>(false);
+  // const [isClient, setIsClient] = useState<boolean>(false);
 
   // forced leave balance
   const [forcedLeaveBalance, setForcedLeaveBalance] = useState<number>(0);
@@ -215,291 +215,279 @@ export const LeaveLedgerPdf: FunctionComponent<LeaveLedgerPdfProps> = ({ employe
   useEffect(() => {
     getLatestBalance(leaveLedger);
 
-    setIsClient(true);
+    // setIsClient(true);
   }, []);
 
   return (
     <>
-      {isClient && (
-        <PDFViewer width={'100%'} height={1400}>
-          <Document title="Leave Ledger">
-            {/* FOLIO */}
-            <Page size={'A4'} orientation={'landscape'} style={styles.page}>
+      {/* {isClient && (
+        <PDFViewer width={'100%'} height={1400}> */}
+      <Document title="Leave Ledger">
+        {/* FOLIO */}
+        <Page size={'A4'} orientation={'landscape'} style={styles.page}>
+          <View>
+            {/* HEADER */}
+            <View style={styles.rowContainer}>
+              {/* LEFT */}
+              <View style={[styles.w30, { padding: '0 0 0 15' }]}>
+                <Image src={GscwdLogo.src} style={[styles.logo]} />
+              </View>
+
+              {/* CENTER */}
+              <View style={[styles.w40, styles.horizontalCenter]}>
+                <Text style={{ fontSize: 11, fontFamily: 'Helvetica-Bold' }}>GENERAL SANTOS WATER DISTRICT</Text>
+                <Text style={{ fontSize: 9, paddingTop: 3 }}>E. Ferdnandez St., Lagao General Santos City</Text>
+                <Text
+                  style={{
+                    fontSize: 11,
+                    paddingTop: 3,
+                    fontFamily: 'Helvetica-Bold',
+                  }}
+                >
+                  LEAVE LEDGER
+                </Text>
+              </View>
+
+              {/* RIGHT */}
+              <View style={[styles.w30]}></View>
+            </View>
+
+            {/* EMPLOYEE PROFILE */}
+            <View style={styles.employeeProfile}>
               <View>
-                {/* HEADER */}
-                <View style={styles.rowContainer}>
-                  {/* LEFT */}
-                  <View style={[styles.w30, { padding: '0 0 0 15' }]}>
-                    <Image src={GscwdLogo.src} style={[styles.logo]} />
-                  </View>
-
-                  {/* CENTER */}
-                  <View style={[styles.w40, styles.horizontalCenter]}>
-                    <Text style={{ fontSize: 11, fontFamily: 'Helvetica-Bold' }}>GENERAL SANTOS WATER DISTRICT</Text>
-                    <Text style={{ fontSize: 9, paddingTop: 3 }}>E. Ferdnandez St., Lagao General Santos City</Text>
-                    <Text
-                      style={{
-                        fontSize: 11,
-                        paddingTop: 3,
+                {/* NAME */}
+                <View style={[styles.rowContainer]}>
+                  <Text style={[styles.w20]}>NAME</Text>
+                  <Text
+                    style={[
+                      styles.w80,
+                      {
                         fontFamily: 'Helvetica-Bold',
-                      }}
-                    >
-                      LEAVE LEDGER
-                    </Text>
-                  </View>
-
-                  {/* RIGHT */}
-                  <View style={[styles.w30]}></View>
+                        textTransform: 'uppercase',
+                        fontSize: 8.3,
+                      },
+                    ]}
+                  >
+                    {employeeData.employmentDetails.employeeFullName}
+                  </Text>
                 </View>
 
-                {/* EMPLOYEE PROFILE */}
-                <View style={styles.employeeProfile}>
-                  <View>
-                    {/* NAME */}
-                    <View style={[styles.rowContainer]}>
-                      <Text style={[styles.w20]}>NAME</Text>
-                      <Text
-                        style={[
-                          styles.w80,
-                          {
-                            fontFamily: 'Helvetica-Bold',
-                            textTransform: 'uppercase',
-                            fontSize: 8.3,
-                          },
-                        ]}
-                      >
-                        {employeeData.employmentDetails.employeeFullName}
-                      </Text>
-                    </View>
-
-                    {/* DESIGNATION */}
-                    <View style={[styles.rowContainer, { paddingTop: 2 }]}>
-                      <Text style={[styles.w20]}>DESIGNATION</Text>
-                      <Text style={[styles.w80, { textTransform: 'uppercase', fontSize: 8 }]}>
-                        {employeeData.employmentDetails.assignment.positionTitle}
-                      </Text>
-                    </View>
-
-                    {/* OFFICE */}
-                    <View style={[styles.rowContainer, { paddingTop: 4 }]}>
-                      <Text style={[styles.w20]}>OFFICE</Text>
-                      <Text style={[styles.w80]}>{employeeData.employmentDetails.orgStruct.officeName}</Text>
-                    </View>
-
-                    {/* DEPARTMENT */}
-                    <View style={[styles.rowContainer]}>
-                      <Text style={[styles.w20]}>DEPARTMENT</Text>
-                      <Text style={[styles.w80]}>{employeeData.employmentDetails.orgStruct.departmentName}</Text>
-                    </View>
-
-                    {/* DIVISION */}
-                    <View style={[styles.rowContainer]}>
-                      <Text style={[styles.w20]}>DIVISION</Text>
-                      <Text style={[styles.w80]}>{employeeData.employmentDetails.orgStruct.divisionName}</Text>
-                    </View>
-                  </View>
+                {/* DESIGNATION */}
+                <View style={[styles.rowContainer, { paddingTop: 2 }]}>
+                  <Text style={[styles.w20]}>DESIGNATION</Text>
+                  <Text style={[styles.w80, { textTransform: 'uppercase', fontSize: 8 }]}>
+                    {employeeData.employmentDetails.assignment.positionTitle}
+                  </Text>
                 </View>
 
-                {/* DAILY TIME LOGS */}
-                <View style={[styles.ledgerContainer]}>
-                  {/* COLUMN HEADERS  */}
-                  <View>
-                    <View style={[styles.rowContainer, styles.rowBorder]}>
-                      <View style={[styles.tableHeader, styles.w8]}>
-                        <Text style={[styles.tableHeaderText]}>PERIOD</Text>
-                      </View>
-                      <View style={[styles.tableHeader, styles.w17]}>
-                        <Text style={[styles.tableHeaderText]}>PARTICULARS</Text>
-                      </View>
-                      <View style={[styles.tableHeader2, styles.w5]}>
-                        <Text style={[styles.tableHeaderText]}>FORCED LEAVE</Text>
-                      </View>
-                      <View style={[styles.tableHeader2, styles.w5]}>
-                        <Text style={[styles.tableHeaderText]}>FORCED LEAVE BAL.</Text>
-                      </View>
-                      <View style={[styles.tableHeader2, styles.w5]}>
-                        <Text style={[styles.tableHeaderText]}>VACATION LEAVE</Text>
-                      </View>
-                      <View style={[styles.tableHeader2, styles.w5]}>
-                        <Text style={[styles.tableHeaderText]}>VACATION LEAVE BAL</Text>
-                      </View>
-                      <View style={[styles.tableHeader2, styles.w5]}>
-                        <Text style={[styles.tableHeaderText]}>SICK LEAVE</Text>
-                      </View>
-                      <View style={[styles.tableHeader2, styles.w5]}>
-                        <Text style={[styles.tableHeaderText]}>SICK LEAVE BAL</Text>
-                      </View>
-                      <View style={[styles.tableHeader2, styles.w5]}>
-                        <Text style={[styles.tableHeaderText]}>SPECIAL PRIVILEGE LEAVE</Text>
-                      </View>
-                      <View style={[styles.tableHeader2, styles.w5]}>
-                        <Text style={[styles.tableHeaderText]}>SPECIAL PRIVILEGE LEAVE BAL</Text>
-                      </View>
-                      <View style={[styles.tableHeader2, styles.w5]}>
-                        <Text style={[styles.tableHeaderText]}>SPECIAL LEAVE BENEFIT</Text>
-                      </View>
-                      <View style={[styles.tableHeader2, styles.w5]}>
-                        <Text style={[styles.tableHeaderText]}>SPECIAL LEAVE BENEFIT BAL</Text>
-                      </View>
-                      <View style={[styles.tableHeader, styles.w25, { borderRight: 'none' }]}>
-                        <Text style={[styles.tableHeaderText]}>REMARKS</Text>
-                      </View>
-                    </View>
+                {/* OFFICE */}
+                <View style={[styles.rowContainer, { paddingTop: 4 }]}>
+                  <Text style={[styles.w20]}>OFFICE</Text>
+                  <Text style={[styles.w80]}>{employeeData.employmentDetails.orgStruct.officeName}</Text>
+                </View>
+
+                {/* DEPARTMENT */}
+                <View style={[styles.rowContainer]}>
+                  <Text style={[styles.w20]}>DEPARTMENT</Text>
+                  <Text style={[styles.w80]}>{employeeData.employmentDetails.orgStruct.departmentName}</Text>
+                </View>
+
+                {/* DIVISION */}
+                <View style={[styles.rowContainer]}>
+                  <Text style={[styles.w20]}>DIVISION</Text>
+                  <Text style={[styles.w80]}>{employeeData.employmentDetails.orgStruct.divisionName}</Text>
+                </View>
+              </View>
+            </View>
+
+            {/* DAILY TIME LOGS */}
+            <View style={[styles.ledgerContainer]}>
+              {/* COLUMN HEADERS  */}
+              <View>
+                <View style={[styles.rowContainer, styles.rowBorder]}>
+                  <View style={[styles.tableHeader, styles.w8]}>
+                    <Text style={[styles.tableHeaderText]}>PERIOD</Text>
                   </View>
-
-                  {/* TABLE ROWS */}
-                  <View>
-                    {!isEmpty(leaveLedger) ? (
-                      leaveLedger.map((entry, index) => {
-                        return (
-                          <View
-                            style={[
-                              styles.rowContainer,
-                              styles.rowBorder,
-                              { backgroundColor: `${leaveBgColor(entry)}` },
-                            ]}
-                            key={index}
-                            wrap={false}
-                          >
-                            <View style={[styles.tableData, styles.w8]}>
-                              <Text style={[styles.tableDataText]}>{dayjs(entry.period).format('MM/DD/YYYY')}</Text>
-                            </View>
-                            <View style={[styles.tableData, styles.w17, { textAlign: 'left' }]}>
-                              <Text style={[styles.tableDataText]}>{entry.particulars}</Text>
-                            </View>
-                            <View style={[styles.tableData, styles.w5]}>
-                              <Text style={[styles.tableDataText]}>
-                                {!isEmpty(entry.forcedLeave) && parseFloat(entry.forcedLeave.toString()) !== 0
-                                  ? entry.forcedLeave
-                                  : ''}
-                              </Text>
-                            </View>
-                            <View style={[styles.tableData, styles.w5]}>
-                              <Text style={[styles.tableDataText]}>{entry.forcedLeaveBalance}</Text>
-                            </View>
-                            <View style={[styles.tableData, styles.w5]}>
-                              <Text style={[styles.tableDataText]}>
-                                {!isEmpty(entry.vacationLeave) && parseFloat(entry.vacationLeave.toString()) !== 0
-                                  ? entry.vacationLeave
-                                  : ''}
-                              </Text>
-                            </View>
-                            <View style={[styles.tableData, styles.w5]}>
-                              <Text style={[styles.tableDataText]}>{entry.vacationLeaveBalance}</Text>
-                            </View>
-                            <View style={[styles.tableData, styles.w5]}>
-                              <Text style={[styles.tableDataText]}>
-                                {!isEmpty(entry.sickLeave) && parseFloat(entry.sickLeave.toString()) !== 0
-                                  ? entry.sickLeave
-                                  : null}
-                              </Text>
-                            </View>
-                            <View style={[styles.tableData, styles.w5]}>
-                              <Text style={[styles.tableDataText]}>{entry.sickLeaveBalance}</Text>
-                            </View>
-                            <View style={[styles.tableData, styles.w5]}>
-                              <Text style={[styles.tableDataText]}>
-                                {!isEmpty(entry.specialPrivilegeLeave) &&
-                                parseFloat(entry.specialPrivilegeLeave.toString()) !== 0
-                                  ? entry.specialPrivilegeLeave
-                                  : null}
-                              </Text>
-                            </View>
-                            <View style={[styles.tableData, styles.w5]}>
-                              <Text style={[styles.tableDataText]}>{entry.specialPrivilegeLeaveBalance}</Text>
-                            </View>
-                            <View style={[styles.tableData, styles.w5]}>
-                              <Text style={[styles.tableDataText]}>
-                                {!isEmpty(entry.specialLeaveBenefit) &&
-                                parseFloat(entry.specialLeaveBenefit.toString()) !== 0
-                                  ? entry.specialLeaveBenefit
-                                  : null}
-                              </Text>
-                            </View>
-                            <View style={[styles.tableData, styles.w5]}>
-                              <Text style={[styles.tableDataText]}>{entry.specialLeaveBenefitBalance}</Text>
-                            </View>
-                            <View style={[styles.tableData, styles.w25, { borderRight: 'none' }]}>
-                              <Text
-                                style={[styles.tableDataText, styles.verticalCenter]}
-                                hyphenationCallback={remarksHyphenationCallback}
-                              >
-                                {entry.remarks} {leaveDatesToString(entry.leaveDates)}
-                              </Text>
-                            </View>
-                          </View>
-                        );
-                      })
-                    ) : (
-                      <View style={[styles.rowContainer, styles.rowBorder]}>
-                        <View style={[styles.tableData, styles.w100]}>
-                          <Text style={[styles.tableDataText]}>NO DATA FOUND</Text>
-                        </View>
-                      </View>
-                    )}
-
-                    {/* TOTAL */}
-                    <View style={[styles.rowContainer, styles.rowBorder]}>
-                      <View style={[styles.tableData, styles.w25]}>
-                        <Text style={[styles.tableDataText, { fontFamily: 'Helvetica-Bold' }]}>TOTAL</Text>
-                      </View>
-                      <View style={[styles.tableData, styles.w5]}>
-                        <Text style={[styles.tableDataText]}></Text>
-                      </View>
-                      <View style={[styles.tableData, styles.w5]}>
-                        <Text
-                          style={[styles.tableDataText, { fontFamily: 'Helvetica-Bold', backgroundColor: '#fecaca' }]}
-                        >
-                          {forcedLeaveBalance}
-                        </Text>
-                      </View>
-                      <View style={[styles.tableData, styles.w5]}>
-                        <Text style={[styles.tableDataText]}></Text>
-                      </View>
-                      <View style={[styles.tableData, styles.w5]}>
-                        <Text
-                          style={[styles.tableDataText, { fontFamily: 'Helvetica-Bold', backgroundColor: '#bbf7d0' }]}
-                        >
-                          {vacationLeaveBalance}
-                        </Text>
-                      </View>
-                      <View style={[styles.tableData, styles.w5]}>
-                        <Text style={[styles.tableDataText]}></Text>
-                      </View>
-                      <View style={[styles.tableData, styles.w5]}>
-                        <Text
-                          style={[styles.tableDataText, { fontFamily: 'Helvetica-Bold', backgroundColor: '#fed7aa' }]}
-                        >
-                          {sickLeaveBalance}
-                        </Text>
-                      </View>
-                      <View style={[styles.tableData, styles.w5]}>
-                        <Text style={[styles.tableDataText]}></Text>
-                      </View>
-                      <View style={[styles.tableData, styles.w5]}>
-                        <Text
-                          style={[styles.tableDataText, { fontFamily: 'Helvetica-Bold', backgroundColor: '#a5f3fc' }]}
-                        >
-                          {specialPrivilegeLeaveBalance}
-                        </Text>
-                      </View>
-                      <View style={[styles.tableData, styles.w5]}>
-                        <Text style={[styles.tableDataText]}></Text>
-                      </View>
-                      <View style={[styles.tableData, styles.w5]}>
-                        <Text style={[styles.tableDataText]}></Text>
-                      </View>
-                      <View style={[styles.tableData, styles.w25, { borderRight: 'none' }]}>
-                        <Text style={[styles.tableDataText]}></Text>
-                      </View>
-                    </View>
+                  <View style={[styles.tableHeader, styles.w17]}>
+                    <Text style={[styles.tableHeaderText]}>PARTICULARS</Text>
+                  </View>
+                  <View style={[styles.tableHeader2, styles.w5]}>
+                    <Text style={[styles.tableHeaderText]}>FORCED LEAVE</Text>
+                  </View>
+                  <View style={[styles.tableHeader2, styles.w5]}>
+                    <Text style={[styles.tableHeaderText]}>FORCED LEAVE BAL.</Text>
+                  </View>
+                  <View style={[styles.tableHeader2, styles.w5]}>
+                    <Text style={[styles.tableHeaderText]}>VACATION LEAVE</Text>
+                  </View>
+                  <View style={[styles.tableHeader2, styles.w5]}>
+                    <Text style={[styles.tableHeaderText]}>VACATION LEAVE BAL</Text>
+                  </View>
+                  <View style={[styles.tableHeader2, styles.w5]}>
+                    <Text style={[styles.tableHeaderText]}>SICK LEAVE</Text>
+                  </View>
+                  <View style={[styles.tableHeader2, styles.w5]}>
+                    <Text style={[styles.tableHeaderText]}>SICK LEAVE BAL</Text>
+                  </View>
+                  <View style={[styles.tableHeader2, styles.w5]}>
+                    <Text style={[styles.tableHeaderText]}>SPECIAL PRIVILEGE LEAVE</Text>
+                  </View>
+                  <View style={[styles.tableHeader2, styles.w5]}>
+                    <Text style={[styles.tableHeaderText]}>SPECIAL PRIVILEGE LEAVE BAL</Text>
+                  </View>
+                  <View style={[styles.tableHeader2, styles.w5]}>
+                    <Text style={[styles.tableHeaderText]}>SPECIAL LEAVE BENEFIT</Text>
+                  </View>
+                  <View style={[styles.tableHeader2, styles.w5]}>
+                    <Text style={[styles.tableHeaderText]}>SPECIAL LEAVE BENEFIT BAL</Text>
+                  </View>
+                  <View style={[styles.tableHeader, styles.w25, { borderRight: 'none' }]}>
+                    <Text style={[styles.tableHeaderText]}>REMARKS</Text>
                   </View>
                 </View>
               </View>
-            </Page>
-          </Document>
-        </PDFViewer>
-      )}
+
+              {/* TABLE ROWS */}
+              <View>
+                {!isEmpty(leaveLedger) ? (
+                  leaveLedger.map((entry, index) => {
+                    return (
+                      <View
+                        style={[styles.rowContainer, styles.rowBorder, { backgroundColor: `${leaveBgColor(entry)}` }]}
+                        key={index}
+                        wrap={false}
+                      >
+                        <View style={[styles.tableData, styles.w8]}>
+                          <Text style={[styles.tableDataText]}>{dayjs(entry.period).format('MM/DD/YYYY')}</Text>
+                        </View>
+                        <View style={[styles.tableData, styles.w17, { textAlign: 'left' }]}>
+                          <Text style={[styles.tableDataText]}>{entry.particulars}</Text>
+                        </View>
+                        <View style={[styles.tableData, styles.w5]}>
+                          <Text style={[styles.tableDataText]}>
+                            {!isEmpty(entry.forcedLeave) && parseFloat(entry.forcedLeave.toString()) !== 0
+                              ? entry.forcedLeave
+                              : ''}
+                          </Text>
+                        </View>
+                        <View style={[styles.tableData, styles.w5]}>
+                          <Text style={[styles.tableDataText]}>{entry.forcedLeaveBalance}</Text>
+                        </View>
+                        <View style={[styles.tableData, styles.w5]}>
+                          <Text style={[styles.tableDataText]}>
+                            {!isEmpty(entry.vacationLeave) && parseFloat(entry.vacationLeave.toString()) !== 0
+                              ? entry.vacationLeave
+                              : ''}
+                          </Text>
+                        </View>
+                        <View style={[styles.tableData, styles.w5]}>
+                          <Text style={[styles.tableDataText]}>{entry.vacationLeaveBalance}</Text>
+                        </View>
+                        <View style={[styles.tableData, styles.w5]}>
+                          <Text style={[styles.tableDataText]}>
+                            {!isEmpty(entry.sickLeave) && parseFloat(entry.sickLeave.toString()) !== 0
+                              ? entry.sickLeave
+                              : null}
+                          </Text>
+                        </View>
+                        <View style={[styles.tableData, styles.w5]}>
+                          <Text style={[styles.tableDataText]}>{entry.sickLeaveBalance}</Text>
+                        </View>
+                        <View style={[styles.tableData, styles.w5]}>
+                          <Text style={[styles.tableDataText]}>
+                            {!isEmpty(entry.specialPrivilegeLeave) &&
+                            parseFloat(entry.specialPrivilegeLeave.toString()) !== 0
+                              ? entry.specialPrivilegeLeave
+                              : null}
+                          </Text>
+                        </View>
+                        <View style={[styles.tableData, styles.w5]}>
+                          <Text style={[styles.tableDataText]}>{entry.specialPrivilegeLeaveBalance}</Text>
+                        </View>
+                        <View style={[styles.tableData, styles.w5]}>
+                          <Text style={[styles.tableDataText]}>
+                            {!isEmpty(entry.specialLeaveBenefit) &&
+                            parseFloat(entry.specialLeaveBenefit.toString()) !== 0
+                              ? entry.specialLeaveBenefit
+                              : null}
+                          </Text>
+                        </View>
+                        <View style={[styles.tableData, styles.w5]}>
+                          <Text style={[styles.tableDataText]}>{entry.specialLeaveBenefitBalance}</Text>
+                        </View>
+                        <View style={[styles.tableData, styles.w25, { borderRight: 'none' }]}>
+                          <Text
+                            style={[styles.tableDataText, styles.verticalCenter]}
+                            hyphenationCallback={remarksHyphenationCallback}
+                          >
+                            {entry.remarks} {leaveDatesToString(entry.leaveDates)}
+                          </Text>
+                        </View>
+                      </View>
+                    );
+                  })
+                ) : (
+                  <View style={[styles.rowContainer, styles.rowBorder]}>
+                    <View style={[styles.tableData, styles.w100]}>
+                      <Text style={[styles.tableDataText]}>NO DATA FOUND</Text>
+                    </View>
+                  </View>
+                )}
+
+                {/* TOTAL */}
+                <View style={[styles.rowContainer, styles.rowBorder]}>
+                  <View style={[styles.tableData, styles.w25]}>
+                    <Text style={[styles.tableDataText, { fontFamily: 'Helvetica-Bold' }]}>TOTAL</Text>
+                  </View>
+                  <View style={[styles.tableData, styles.w5]}>
+                    <Text style={[styles.tableDataText]}></Text>
+                  </View>
+                  <View style={[styles.tableData, styles.w5]}>
+                    <Text style={[styles.tableDataText, { fontFamily: 'Helvetica-Bold', backgroundColor: '#fecaca' }]}>
+                      {forcedLeaveBalance}
+                    </Text>
+                  </View>
+                  <View style={[styles.tableData, styles.w5]}>
+                    <Text style={[styles.tableDataText]}></Text>
+                  </View>
+                  <View style={[styles.tableData, styles.w5]}>
+                    <Text style={[styles.tableDataText, { fontFamily: 'Helvetica-Bold', backgroundColor: '#bbf7d0' }]}>
+                      {vacationLeaveBalance}
+                    </Text>
+                  </View>
+                  <View style={[styles.tableData, styles.w5]}>
+                    <Text style={[styles.tableDataText]}></Text>
+                  </View>
+                  <View style={[styles.tableData, styles.w5]}>
+                    <Text style={[styles.tableDataText, { fontFamily: 'Helvetica-Bold', backgroundColor: '#fed7aa' }]}>
+                      {sickLeaveBalance}
+                    </Text>
+                  </View>
+                  <View style={[styles.tableData, styles.w5]}>
+                    <Text style={[styles.tableDataText]}></Text>
+                  </View>
+                  <View style={[styles.tableData, styles.w5]}>
+                    <Text style={[styles.tableDataText, { fontFamily: 'Helvetica-Bold', backgroundColor: '#a5f3fc' }]}>
+                      {specialPrivilegeLeaveBalance}
+                    </Text>
+                  </View>
+                  <View style={[styles.tableData, styles.w5]}>
+                    <Text style={[styles.tableDataText]}></Text>
+                  </View>
+                  <View style={[styles.tableData, styles.w5]}>
+                    <Text style={[styles.tableDataText]}></Text>
+                  </View>
+                  <View style={[styles.tableData, styles.w25, { borderRight: 'none' }]}>
+                    <Text style={[styles.tableDataText]}></Text>
+                  </View>
+                </View>
+              </View>
+            </View>
+          </View>
+        </Page>
+      </Document>
+      {/* </PDFViewer>
+      )} */}
     </>
   );
 };
